legal-ai-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
openai==1.3.7
anthropic==0.7.8
python-dotenv==1.0.0
aiofiles==23.2.1
pymupdf==1.23.8
langchain==0.0.350
langchain-community==0.0.1